Extract Roam importer into a plugin
src/utils/roamImport/ → src/plugins/roam-import/ (all 13 source
modules + 6 test files moved via git mv to preserve history). The
old action handler and the window-hook bootstrap are out of
defaultShortcuts.ts; the plugin now owns both.

- `roamImportPlugin({repo})` contributes `import_roam` to
  actionsFacet and `roam-import.window-hook` to appEffectsFacet.
  The AppEffect installs the idempotent
  `window.__omniliner.roamImport` debug hook once per Repo — same
  surface that used to live inline at the top of
  getDefaultActionGroups.
- defaultShortcuts.ts loses the import_roam action body, the
  ensureRoamImportWindowHook call, and four Roam-specific imports.
  The ensureMetricsConsoleHook bootstrap stays — it's not Roam-
  shaped and there's no plugin home for it yet.
- progressBanner moves to src/utils/progressBanner.ts. Both the
  Roam importer action and the SQLite export/import actions (still
  in defaultShortcuts.ts) consume it; the lint-cleanest spot for a
  shared UI helper is /utils/. When SQLite export gets its own
  plugin (next on the audit list), this can move with one of them.
- src/plugins/roam-import/index.ts gains a `roamImportPlugin`
  export alongside the pre-existing `importRoam`/`planImport`/etc.
  public surface — the importer's API stays available to agent-
  runtime scripts and to anything else that wants to drive it
  programmatically.
- Internal `../../data/...` paths inside roamImport rewrite to
  `@/data/...` (only two leaks: import.ts and runtime.ts). The
  rest were already alias-qualified.
- Test fixture path probes in sampleExport.test.ts are unchanged
  (they resolve from process.cwd(), not from the file).
- The lint-rule fixture filename in
  blockSubscriptionLintRules.test.ts updates to the new path.
