+++ src/utils/progressBanner.ts
